// frontend/src/core/api/generated/model/companyProfileView.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Constructor ERP API
 * OpenAPI spec version: 1.0
 */

export interface CompanyProfileView {
  id: string
  name: string
  slug: string
  /** @nullable */
  phone?: string | null
  /** @nullable */
  email?: string | null
  /** @nullable */
  city?: string | null
  /** @nullable */
  logoUrl?: string | null
}